move distance function
